Encryption of files and text 1.0
